move fileifo type out, expose getGameFiles.tsx
